:white_check_mark: Fix test related to discovery
